aw-server: Make UUID part of ServerState
