+++ aw-server/src/endpoints/mod.rs
@@ -1,4 +1,3 @@
-use std::fs;
 use std::path::PathBuf;
 use std::sync::Mutex;
 
@@ -6,10 +5,8 @@ use gethostname::gethostname;
 use rocket::response::NamedFile;
 use rocket::State;
 use rocket_contrib::json::JsonValue;
-use uuid::Uuid;
 
 use crate::config::AWConfig;
-use crate::dirs;
 
 #[macro_export]
 macro_rules! endpoints_get_lock {
@@ -36,6 +33,7 @@ use aw_datastore::Datastore;
 pub struct ServerState {
     pub datastore: Mutex<Datastore>,
     pub asset_path: PathBuf,
+    pub device_id: String,
 }
 
 #[get("/")]
@@ -68,25 +66,8 @@ fn root_favicon(state: State<ServerState>) -> Option<NamedFile> {
     NamedFile::open(state.asset_path.join("favicon.ico")).ok()
 }
 
-/// Retrieves the device ID, if none exists it generates one (using UUID v4)
-fn get_device_id() -> String {
-    // TODO: Cache to avoid retrieving on every /info call
-    // TODO: How should these unwraps be removed?
-    //       Should this be propagated into a 500 Internal Server Error? How?
-    // I chose get_data_dir over get_config_dir since the latter isn't yet supported on Android.
-    let mut path = dirs::get_data_dir().unwrap();
-    path.push("device_id");
-    if path.exists() {
-        fs::read_to_string(path).unwrap()
-    } else {
-        let uuid = Uuid::new_v4().to_hyphenated().to_string();
-        fs::write(path, &uuid).unwrap();
-        uuid
-    }
-}
-
 #[get("/")]
-fn server_info(config: State<AWConfig>) -> JsonValue {
+fn server_info(config: State<AWConfig>, state: State<ServerState>) -> JsonValue {
     #[allow(clippy::or_fun_call)]
     let hostname = gethostname().into_string().unwrap_or("unknown".to_string());
     const VERSION: Option<&'static str> = option_env!("CARGO_PKG_VERSION");
@@ -95,7 +76,7 @@ fn server_info(config: State<AWConfig>) -> JsonValue {
         "hostname": hostname,
         "version": format!("v{} (rust)", VERSION.unwrap_or("(unknown)")),
         "testing": config.testing,
-        "device_id": get_device_id(),
+        "device_id": state.device_id,
     })
 }
